requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
structlog==25.2.0
tomli==2.2.1
typing_extensions==4.13.1
uritemplate==4.1.1
